Generlize the hook
