+++ src/QuickSignPDF/QuickSignPDF.tsx
@@ -33,6 +33,7 @@ import ColorPickerButton, {
   rgbColorToCssRgba,
 } from '../ui/ColorPickerButton';
 import ImagePickerButton from '../ui/ImagePickerButton';
+import useCanvasKeyBindings from '../useCanvasKeybindings';
 
 export interface Signature {
   svg: JSX.Element;
@@ -316,6 +317,7 @@ const QuickSignPDF = (): JSX.Element => {
   );
 
   const editorRef = useRef<FabricJSEditor>(null);
+  useCanvasKeyBindings(editorRef.current);
 
   const updatePageObjects = useCallback(() => {
     setPagesFabricObjects((currentPagesObjects) => {
@@ -417,18 +419,21 @@ const QuickSignPDF = (): JSX.Element => {
     const selectedObjects = editor.canvas.getActiveObjects();
     if (!selectedObjects || selectedObjects.length === 0) return;
 
+    editor.canvas.discardActiveObject();
     const json = selectedObjects.map((object: any) => object.toJSON());
 
     // Group to get the preview data uri. Ungroup right away because there's a bug
     // with Fabric where the selection box act wonky after grouping.
-    const selectedGroup = new fabric.fabric.Group(selectedObjects, {
-      originX: 'center',
-      originY: 'center',
-    });
+    const selectedGroup = new fabric.fabric.Group(selectedObjects);
     const image = selectedGroup.toDataURL({
       format: 'png',
     });
+
     selectedGroup.ungroupOnCanvas();
+    for (const selectedObj of selectedGroup.getObjects()) {
+      editor.canvas.remove(selectedObj);
+      editor.canvas.add(selectedObj);
+    }
 
     const selectedSignature = {
       json,
